requirements.txt
pyspark==3.5.6
kafka-python==2.0.2